# docs/openapi.json
{
  "openapi" : "3.0.1",
  "info" : {
    "title" : "API Documentation",
    "version" : "1.0"
  },
  "paths" : {
...
    "/auctions" : {
      "get" : {
        "summary" : "경매 목록을 페이징 조회합니다.",
        "description" : "경매 목록을 페이징 조회합니다.",
        "parameters" : [ {
          "name" : "status",
          "in" : "query",
          "description" : "경매 상태 필터 (선택). null이면 전체 조회 SCHEDULED :SCHEDULED\nACTIVE :ACTIVE\nENDED :ENDED\nCANCELLED :CANCELLED",
          "required" : false,
          "schema" : {
            "type" : "string",
            "nullable" : false
          }
        } ],
        "requestBody" : {
          "content" : {
            "multipart/form-data" : {
              "schema" : {
                "type" : "object",
                "properties" : {
                  "page" : {
                    "type" : "integer",
                    "nullable" : false
                  },
                  "size" : {
                    "type" : "integer",
                    "nullable" : false
                  },
                  "sort" : {
                    "type" : "array",
                    "nullable" : false,
                    "items" : {
                      "type" : "string"
                    }
                  }
                }
              }
            }
          },
          "required" : false
        },
        "responses" : {
          "200" : {
            "description" : "경매 요약 목록 (페이징)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "content" : {
                      "type" : "array",
                      "description" : "현재 페이지의 데이터 목록",
                      "items" : {
                        "type" : "object",
                        "properties" : {
                          "id" : {
                            "type" : "integer",
                            "description" : "경매 ID",
                            "format" : "int64"
                          },
                          "productId" : {
                            "type" : "integer",
                            "description" : "상품 ID (product-service 참조)",
                            "format" : "int64"
                          },
                          "productName" : {
                            "type" : "string",
                            "description" : "상품명"
                          },
                          "productThumbnailUrl" : {
                            "type" : "string",
                            "description" : "상품 썸네일 URL"
                          },
                          "sellerNickname" : {
                            "type" : "string",
                            "description" : "판매자 닉네임"
                          },
                          "currentPrice" : {
                            "type" : "number",
                            "description" : "현재 최고 입찰가"
                          },
                          "auctionStatus" : {
                            "type" : "string",
                            "description" : "{SCHEDULED=SCHEDULED, ACTIVE=ACTIVE, ENDED=ENDED, CANCELLED=CANCELLED}",
                            "enum" : [ "SCHEDULED", "ACTIVE", "ENDED", "CANCELLED" ]
                          },
                          "endTime" : {
                            "type" : "string",
                            "description" : "경매 종료 시각"
                          },
                          "createdAt" : {
                            "type" : "string",
                            "description" : "경매 등록 일시"
                          }
                        },
                        "description" : "경매 목록 응답 DTO.\n\n<p>목록 조회 시 사용되는 요약 정보입니다.\n상세 정보가 필요한 경우{@code GET /auctions/{id}}를 사용하세요."
                      }
                    },
                    "page" : {
                      "type" : "integer",
                      "description" : "현재 페이지 번호 (0-based)"
                    },
                    "size" : {
                      "type" : "integer",
                      "description" : "페이지 크기"
                    },
                    "totalElements" : {
                      "type" : "integer",
                      "description" : "전체 데이터 건수",
                      "format" : "int64"
                    },
                    "totalPages" : {
                      "type" : "integer",
                      "description" : "전체 페이지 수"
                    },
                    "first" : {
                      "type" : "boolean",
                      "description" : "첫 번째 페이지 여부"
                    },
                    "last" : {
                      "type" : "boolean",
                      "description" : "마지막 페이지 여부"
                    }
                  },
                  "description" : "경매 요약 목록 (페이징)"
                }
              }
            }
          }
        }
      },
      "post" : {
        "summary" : "새 경매를 생성합니다.",
        "description" : "새 경매를 생성합니다.",
        "parameters" : [ {
          "name" : "nickname",
          "in" : "header",
          "description" : "판매자 닉네임 (요청 헤더)",
          "required" : true,
          "schema" : { }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "required" : [ "endTime", "minimumBidIncrement", "productId", "startTime", "startingPrice" ],
                "type" : "object",
                "properties" : {
                  "productId" : {
                    "type" : "integer",
                    "description" : "경매 대상 상품 ID (product-service 참조)",
                    "format" : "int64"
                  },
                  "startingPrice" : {
                    "minimum" : 0,
                    "exclusiveMinimum" : false,
                    "type" : "number",
                    "description" : "시작가. 첫 입찰 시 이 금액 이상이어야 합니다."
                  },
                  "buyNowPrice" : {
                    "minimum" : 0,
                    "exclusiveMinimum" : false,
                    "type" : "number",
                    "description" : "즉시 구매가 (선택). 이 금액 이상의 입찰이 들어오면 경매가 즉시 종료됩니다."
                  },
                  "minimumBidIncrement" : {
                    "minimum" : 0,
                    "exclusiveMinimum" : false,
                    "type" : "number",
                    "description" : "최소 입찰 단위. 현재가에 이 금액을 더한 값 이상으로 입찰해야 합니다."
                  },
                  "startTime" : {
                    "type" : "string",
                    "description" : "경매 시작 시각. 반드시 미래 시각이어야 합니다."
                  },
                  "endTime" : {
                    "type" : "string",
                    "description" : "경매 종료 시각. 반드시 {@code startTime}보다 늦어야 합니다."
                  }
                },
                "description" : "경매 생성 요청 데이터"
              }
            }
          }
        },
        "responses" : {
          "200" : {
            "description" : "생성된 경매 상세 정보 (HTTP 201)",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "id" : {
                      "type" : "integer",
                      "description" : "경매 ID",
                      "format" : "int64"
                    },
                    "productId" : {
                      "type" : "integer",
                      "description" : "상품 ID (product-service 참조)",
                      "format" : "int64"
                    },
                    "productName" : {
                      "type" : "string",
                      "description" : "상품명 (경매 생성 시점 스냅샷)"
                    },
                    "productThumbnailUrl" : {
                      "type" : "string",
                      "description" : "상품 썸네일 URL (경매 생성 시점 스냅샷, 없을 수 있음)"
                    },
                    "sellerNickname" : {
                      "type" : "string",
                      "description" : "판매자 닉네임"
                    },
                    "startingPrice" : {
                      "type" : "number",
                      "description" : "시작가"
                    },
                    "currentPrice" : {
                      "type" : "number",
                      "description" : "현재 최고 입찰가"
                    },
                    "buyNowPrice" : {
                      "type" : "number",
                      "description" : "즉시 구매가 (설정하지 않은 경우 null)"
                    },
                    "minimumBidIncrement" : {
                      "type" : "number",
                      "description" : "최소 입찰 단위"
                    },
                    "auctionStatus" : {
                      "type" : "string",
                      "description" : "{SCHEDULED=SCHEDULED, ACTIVE=ACTIVE, ENDED=ENDED, CANCELLED=CANCELLED}",
                      "enum" : [ "SCHEDULED", "ACTIVE", "ENDED", "CANCELLED" ]
                    },
                    "startTime" : {
                      "type" : "string",
                      "description" : "경매 시작 시각"
                    },
                    "endTime" : {
                      "type" : "string",
                      "description" : "경매 종료 시각"
                    },
                    "winnerNickname" : {
                      "type" : "string",
                      "description" : "낙찰자 닉네임 (경매 종료 전 null)"
                    },
                    "finalPrice" : {
                      "type" : "number",
                      "description" : "낙찰가 (경매 종료 전 null)"
                    },
                    "createdAt" : {
                      "type" : "string",
                      "description" : "경매 등록 일시"
                    },
                    "updatedAt" : {
                      "type" : "string",
                      "description" : "경매 최종 수정 일시"
                    }
                  },
                  "description" : "생성된 경매 상세 정보 (HTTP 201)"
                }
              }
            }
          }
        }
      }
    },
    "/auctions/{auctionId}" : {
      "get" : {
        "summary" : "경매 상세 정보를 조회합니다.",
        "description" : "경매 상세 정보를 조회합니다.",
        "parameters" : [ {
          "name" : "auctionId",
          "in" : "path",
          "description" : "경매 ID",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "경매 상세 정보",
            "content" : {
              "application/json" : {
                "schema" : {
                  "type" : "object",
                  "properties" : {
                    "id" : {
                      "type" : "integer",
                      "description" : "경매 ID",
                      "format" : "int64"
                    },
                    "productId" : {
                      "type" : "integer",
                      "description" : "상품 ID (product-service 참조)",
                      "format" : "int64"
                    },
                    "productName" : {
                      "type" : "string",
                      "description" : "상품명 (경매 생성 시점 스냅샷)"
                    },
                    "productThumbnailUrl" : {
                      "type" : "string",
                      "description" : "상품 썸네일 URL (경매 생성 시점 스냅샷, 없을 수 있음)"
                    },
                    "sellerNickname" : {
                      "type" : "string",
                      "description" : "판매자 닉네임"
                    },
                    "startingPrice" : {
                      "type" : "number",
                      "description" : "시작가"
                    },
                    "currentPrice" : {
                      "type" : "number",
                      "description" : "현재 최고 입찰가"
                    },
                    "buyNowPrice" : {
                      "type" : "number",
                      "description" : "즉시 구매가 (설정하지 않은 경우 null)"
                    },
                    "minimumBidIncrement" : {
                      "type" : "number",
                      "description" : "최소 입찰 단위"
                    },
                    "auctionStatus" : {
                      "type" : "string",
                      "description" : "{SCHEDULED=SCHEDULED, ACTIVE=ACTIVE, ENDED=ENDED, CANCELLED=CANCELLED}",
                      "enum" : [ "SCHEDULED", "ACTIVE", "ENDED", "CANCELLED" ]
                    },
                    "startTime" : {
                      "type" : "string",
                      "description" : "경매 시작 시각"
                    },
                    "endTime" : {
                      "type" : "string",
                      "description" : "경매 종료 시각"
                    },
                    "winnerNickname" : {
                      "type" : "string",
                      "description" : "낙찰자 닉네임 (경매 종료 전 null)"
                    },
                    "finalPrice" : {
                      "type" : "number",
                      "description" : "낙찰가 (경매 종료 전 null)"
                    },
                    "createdAt" : {
                      "type" : "string",
                      "description" : "경매 등록 일시"
                    },
                    "updatedAt" : {
                      "type" : "string",
                      "description" : "경매 최종 수정 일시"
                    }
                  },
                  "description" : "경매 상세 정보"
                }
              }
            }
          }
        }
      },
      "delete" : {
        "summary" : "경매를 취소합니다.",
        "description" : "경매를 취소합니다.",
        "parameters" : [ {
          "name" : "auctionId",
          "in" : "path",
          "description" : "취소할 경매 ID",
          "required" : true,
          "schema" : {
            "type" : "integer"
          }
        }, {
          "name" : "nickname",
          "in" : "header",
          "description" : "판매자 닉네임 (요청 헤더)",
          "required" : true,
          "schema" : { }
        } ],
        "responses" : {
          "200" : {
            "description" : "No Content"
          }
        }
      }
    },
...
  },
...
}